chore(qa): bump PHPStan to level 7, fix type annotations, regenerate testing report
Fix all 96 level-7 type errors across 20 source files. Patterns addressed:
property.nonObject (WP_Query->posts int|WP_Post), property.notFound (object
vs stdClass for DB rows), argument.type (strtotime int|false, fgetcsv
resource|false, WP_Error::get_error_code int|string), return.type (apply_filters
mixed, shape-typed arrays), method.notFound (WC_Order vs WC_Order_Refund),
foreach.nonIterable (wc_get_orders paginated stdClass). Also sets
maximumNumberOfProcesses:1 to prevent OOM from parallel WooCommerce stubs load.

PHPStan: PASS 0 errors | Unit tests: 276 tests, 911 assertions, 0 failures | PHPCS: 0 errors

+++ wp-content/plugins/gym-core/src/Gamification/StreakTracker.php
@@ -130,7 +130,7 @@ class StreakTracker {
 		$previous_week = $weeks[0];
 
 		for ( $i = 1, $count = count( $weeks ); $i < $count; $i++ ) {
-			$expected_next = gmdate( 'Y-m-d', strtotime( $previous_week . ' +7 days' ) );
+			$expected_next = gmdate( 'Y-m-d', (int) strtotime( $previous_week . ' +7 days' ) );
 
 			if ( $weeks[ $i ] === $expected_next ) {
 				// Consecutive week — extend streak.
